Remove key hashing and move to salus-commmon
